Refactor EventD to clean up the rpc servicer a bit (#4184)
